refactor: extract helper utilities
